Split installation and provisioning into separate builds
The provisioner doesn't know about the installation process we're going
through, rather it only knows that it's run the boot command and is now
waiting for SSH to become available.

This leads to a problem where it attempts to provision the vagrant user
onto the live server since it has a running SSH server, resulting in the
output box not having this user.

There are some hacks and workaround that can be employed to force the
provisioner to simply wait a given amount of time before trying to
provision the box. This is unreliable since we don't know how long the
initial installation will take to complete.

The robust solution is to create a build pipeline where the installation
takes place in the first stage and a VMWare VMX file is produced without
the vagrant user. This VMX is fed into the second build step where the
vagrant user is provisioned and a Vagrant box is produced.